feat: implement ConfigurationType for JSON serialization of Configuration subclasses
