+++ edge_mining/adapters/domain/home_load/tables.py
@@ -13,6 +13,7 @@ from typing import Optional
 
 from sqlalchemy import JSON, Column, String, Table, TypeDecorator, event
 
+from edge_mining.adapters.infrastructure.persistence.sqlalchemy.common import ConfigurationType
 from edge_mining.adapters.infrastructure.persistence.sqlalchemy.registry import mapper_registry, metadata
 from edge_mining.domain.common import EntityId
 from edge_mining.domain.home_load.aggregate_roots import HomeLoadsProfile
@@ -23,27 +24,11 @@ from edge_mining.shared.adapter_maps.home_load import HOME_FORECAST_PROVIDER_CON
 from edge_mining.shared.interfaces.config import HomeForecastProviderConfig
 
 
-class HomeForecastProviderConfigType(TypeDecorator):
-    """Custom SQLAlchemy type that converts HomeForecastProviderConfig to/from JSON string.
+class HomeForecastProviderConfigType(ConfigurationType):
+    """SQLAlchemy type for HomeForecastProviderConfig serialization.
 
-    This type handles serialization when writing to the database.
-    Deserialization is handled by the @event.listens_for decorator on the entity.
+    Inherits from ConfigurationType to handle JSON serialization/deserialization.
     """
-
-    impl = String
-    cache_ok = True
-
-    def process_bind_param(self, value: Optional[HomeForecastProviderConfig], dialect) -> Optional[str]:
-        """Convert HomeForecastProviderConfig to JSON string before storing in DB."""
-        if value is None:
-            return None
-        if isinstance(value, str):
-            return value
-        return json.dumps(value.to_dict())
-
-    def process_result_value(self, value: Optional[str], dialect) -> Optional[str]:
-        """Return the JSON string as-is. Actual deserialization happens in the event listener."""
-        return value
 
 
 def _deserialize_home_forecast_provider_config(
